worked_maml_ad

+++ utils.py
@@ -99,7 +99,7 @@ def get_data_from_source(args, normalize=False):
     print("train set shape: ", train_data.shape)
     print("test set shape: ", test_data.shape)
     print("test set label shape: ", None if test_label is None else test_label.shape)
-    train_label = np.zeros(len(train_data))
+    train_label = np.zeros(len(train_data),dtype=np.float32)
     return (train_data, train_label), (test_data, test_label)
